philo : change starting time

+++ philosophers/includes/philosophers.h
@@ -17,7 +17,7 @@ typedef struct s_params
     int             to_sleep;
     int             philos;
     int             eat_max;
-    int             start;
+    long int        start;
     int             died;
     pthread_mutex_t print;
 }               t_params;
@@ -25,7 +25,7 @@ typedef struct s_params
 typedef struct s_phi
 {
     int             id;
-    int             start_eating;
+    long int        start_eating;
     pthread_mutex_t *l_fork;
     pthread_mutex_t *r_fork;
     pthread_t		thread;
